[NFC] Refactor ThinBackend
 - Change it to a type from a function.
 - Store the parallelism in the type for the future use.
